Feature/first choose flower (#4)
* feat: apiSetting (#1)

* feat: addchooseflower

* fix: delete

---------
